Moves afq_profile and related tests to dipy.stats.analysis.
